Reorganizing module namespace order
The previous reasoning was faulty, and I should have talked it out.
I was thinking it would be better to have a named module that
contained the necessary submodules.

However, as I was exploring how to reconcile that with view paths and
rendering options, it became clear that I needed to put the concept
(i.e. WorkSubmissions or SubmisionWindows) as the container of the
work area (i.e. Etd).
