--- app/forms/sipity/forms/core.rb
@@ -1,11 +1,0 @@
-module Sipity
-  module Forms
-    # A container for classes that are core to Sipity; That is to say things
-    # I don't believe to be part of a potential plugin.
-    #
-    # Why not Base? I believe Base to be more representative of things you build
-    # from, not stand alone elements.
-    module Core
-    end
-  end
-end
