--- app/forms/sipity/forms/ulra.rb
@@ -1,7 +1,0 @@
-module Sipity
-  module Forms
-    # The container module for all ULRA related forms.
-    module Ulra
-    end
-  end
-end
