--- app/forms/sipity/forms/ulra/submission_windows.rb
@@ -1,9 +1,0 @@
-module Sipity
-  module Forms
-    module Ulra
-      # Container module for SubmissionWindows within ULRA
-      module SubmissionWindows
-      end
-    end
-  end
-end
